requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
opentelemetry-api==1.33.1
opentelemetry-distro==0.54b1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-asyncio==0.54b1
opentelemetry-instrumentation-click==0.54b1
opentelemetry-instrumentation-dbapi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-instrumentation-grpc==0.54b1
opentelemetry-instrumentation-httpx==0.54b1
opentelemetry-instrumentation-jinja2==0.54b1
opentelemetry-instrumentation-logging==0.54b1
opentelemetry-instrumentation-requests==0.54b1
opentelemetry-instrumentation-sqlite3==0.54b1
opentelemetry-instrumentation-starlette==0.54b1
opentelemetry-instrumentation-threading==0.54b1
opentelemetry-instrumentation-tortoiseorm==0.54b1
opentelemetry-instrumentation-urllib==0.54b1
opentelemetry-instrumentation-urllib3==0.54b1
opentelemetry-instrumentation-wsgi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
packaging==25.0
protobuf==5.29.4
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.6
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
zipp==3.22.0
